+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 0.0.6
 
+ - World Clock screen
+ - Changed syslog default protocol to UDP
  - Updates for KFCWebBuilder (requires PHP 7.4.28)
  - Updated weather station plugin for Arduino Core 3.x
  - Fixed memory leak in GFXCanvas